cleanup: rename root source files of "codec" component
