+++ src/codec/CMakeLists.txt
@@ -1,11 +1,11 @@
 # Copyright (c) 2012-2020 Esri R&D Center Zurich
- 
+
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
- 
+
 #   http://www.apache.org/licenses/LICENSE-2.0
- 
+
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,7 +20,7 @@ cmake_policy(SET CMP0015 NEW)
 ### setup build target
 
 add_library(${CODEC_TARGET} SHARED
-		codecs_Py.cpp
+		codec.cpp
 		encoder/PyEncoder.cpp)
 
 target_include_directories(${CODEC_TARGET} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
