+++ src/codec/encoder/IPyCallbacks.h
@@ -19,7 +19,7 @@
 
 #pragma once
 
-#include "codecs_Py.h"
+#include "codec.h"
 #include "prt/Callbacks.h"
